Project 2/requirements.txt
streamlit==1.36.0
langchain==0.2.5
openai==1.35.3
streamlit-chat==0.1.1
python-dotenv==1.0.1
langchain-openai==0.1.9
